requirements.txt
REQUIREMENTS_CONTENT = '''# Core dependencies
google-generativeai>=0.3.0
playwright>=1.40.0
beautifulsoup4>=4.12.0
pandas>=2.0.0
aiohttp>=3.9.0

# Async support
nest-asyncio>=1.5.8

# Environment management
python-dotenv>=1.0.0

# Logging
colorlog>=6.8.0

# Development dependencies (optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
isort>=5.12.0
pylint>=3.0.0
'''